Added Role class, refactored Role based code
Added Role class, refactored Role based code
