+++ dynamo/role.json
@@ -2,7 +2,7 @@
 	"TableName": "Role",
 	"Item": {
 	    "RoleID": {"S": "c104ea59-7deb-4ae4-8418-225d8f4f42cd"},
-        "RoleType": {"S": "admin"},
+        "RoleType": {"S": "Admin"},
     	"RoleName": {"S": "Admin"},
     	"Permissions": {"M": 
     		{ "User_CanCreate": { "N": "1"},
