+++ dynamo/user.json
@@ -5,10 +5,7 @@
 		"Username": {"S": "Admin"},
 		"Password": {"S": "password123"},
 		"Email": {"S": "email@gmail.com"},
-		"UserRoles": {"L": [
-				{"S": "c104ea59-7deb-4ae4-8418-225d8f4f42cd"}
-			]
-		}
+		"Role": {"S": "c104ea59-7deb-4ae4-8418-225d8f4f42cd"}
 	},
 	"ReturnConsumedCapacity": "TOTAL"
 }
